fix discover movie

+++ app/src/main/java/com/qoolqas/moviecatalogue/pojo/movie/MovieData.java
@@ -37,7 +37,7 @@ public class MovieData {
 	private double popularity;
 
 	@SerializedName("vote_average")
-	private int voteAverage;
+	private double voteAverage;
 
 	@SerializedName("id")
 	private int id;
@@ -132,7 +132,7 @@ public class MovieData {
 		this.voteAverage = voteAverage;
 	}
 
-	public int getVoteAverage(){
+	public double getVoteAverage() {
 		return voteAverage;
 	}
